pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pp-boatwatch"
version = "0.0.1"
dependencies = [
  "aiohttp==3.9.3",
  "aiosignal==1.3.1",
  "attrs==23.2.0",
  "certifi==2024.2.2",
  "charset-normalizer==3.3.2",
  "filelock==3.13.1",
  "frozenlist==1.4.1",
  "fsspec==2024.2.0",
  "huggingface-hub==0.20.3",
  "idna==3.6",
  "Jinja2==3.1.3",
  "MarkupSafe==2.1.5",
  "mpmath==1.3.0",
  "multidict==6.0.5",
  "networkx==3.2.1",
  "numpy==1.26.4",
  "nvidia-cublas-cu12==12.1.3.1",
  "nvidia-cuda-cupti-cu12==12.1.105",
  "nvidia-cuda-nvrtc-cu12==12.1.105",
  "nvidia-cuda-runtime-cu12==12.1.105",
  "nvidia-cudnn-cu12==8.9.2.26",
  "nvidia-cufft-cu12==11.0.2.54",
  "nvidia-curand-cu12==10.3.2.106",
  "nvidia-cusolver-cu12==11.4.5.107",
  "nvidia-cusparse-cu12==12.1.0.106",
  "nvidia-nccl-cu12==2.19.3",
  "nvidia-nvjitlink-cu12==12.3.101",
  "nvidia-nvtx-cu12==12.1.105",
  "packaging==23.2",
  "pillow==10.2.0",
  "PyYAML==6.0.1",
  "regex==2023.12.25",
  "requests==2.31.0",
  "safetensors==0.4.2",
  "sympy==1.12",
  "tokenizers==0.15.1",
  "torch==2.2.0",
  "tqdm==4.66.1",
  "transformers==4.37.2",
  "triton==2.2.0",
  "typing_extensions==4.9.0",
  "urllib3==2.2.0",
  "yarl==1.9.4",
]

[project.scripts]
find-objects = "ppboatwatch.find_objects:main"
sample-streams = "ppboatwatch.sample_streams:main"

[tool.setuptools]
packages = ["ppboatwatch"]
